refactor: add view model concept to components
