+++ apps/web/components/ControlPanel/ControlPanel.tsx
@@ -1,4 +1,3 @@
-import { Artboard, ShapeLayer, TextLayer, useUIData } from '@pigyuma/ui-design-tool';
 import clsx from 'clsx';
 import React from 'react';
 import LayerFillSection from '../LayerFillSection';
@@ -8,20 +7,15 @@ import LayerTypographySection from '../LayerTypographySection';
 import Panel from '../Panel';
 import * as styles from './ControlPanel.css';
 import { ControlPanelProps, ControlPanelRef } from './types';
+import useControlPanel from './useControlPanel';
 
 const ControlPanel = React.forwardRef<ControlPanelRef, ControlPanelProps>((props, ref) => {
-  const uiData = useUIData();
-  const selectedRecord = uiData.get([...uiData.selected][0]);
+  const viewModel = useControlPanel(props, ref);
+  if (viewModel == null) {
+    return null;
+  }
 
-  const hasUIRecord = selectedRecord != null;
-
-  const isArtboard = Artboard.isModel(selectedRecord);
-  const isShapeLayer = ShapeLayer.isModel(selectedRecord);
-  const isTextLayer = TextLayer.isModel(selectedRecord);
-
-  const canEditFill = hasUIRecord && (isArtboard || isShapeLayer);
-  const canEditStroke = hasUIRecord && isShapeLayer;
-  const canEditTypography = hasUIRecord && isTextLayer;
+  const { selectedRecord, canEditFill, canEditStroke, canEditTypography } = viewModel;
 
   return (
     <Panel {...props} ref={ref} as="div" role="toolbar" className={clsx(styles.root, props.className)} placement="right">
