+++ apps/web/components/LayerStrokeSection/LayerStrokeSection.tsx
@@ -1,9 +1,6 @@
 import ColorPicker from '@pigyuma/design-system/components/ColorPicker';
 import NumberField from '@pigyuma/design-system/components/NumberField';
 import Select from '@pigyuma/design-system/components/Select';
-import { useEvent } from '@pigyuma/react-utils';
-import { Artboard, ShapeLayer, TextLayer, useUIController, useUIRecord } from '@pigyuma/ui-design-tool';
-import { StrokeStylePattern } from '@pigyuma/ui-design-tool/types/Unit';
 import clsx from 'clsx';
 import React from 'react';
 import LayerField from '../LayerField';
@@ -11,72 +8,30 @@ import LayerFieldset from '../LayerFieldset';
 import Panel from '../Panel';
 import * as styles from './LayerStrokeSection.css';
 import { LayerStrokeSectionProps, LayerStrokeSectionRef } from './types';
+import useLayerStrokeSection from './useLayerStrokeSection';
 
 const LayerStrokeSection = React.forwardRef<LayerStrokeSectionRef, LayerStrokeSectionProps>((props, ref) => {
-  const { selected: selectedRecordKey } = props;
+  const viewModel = useLayerStrokeSection(props, ref);
+  if (viewModel == null) {
+    return null;
+  }
 
-  const uiController = useUIController();
-
-  const uiRecord = useUIRecord<Artboard | ShapeLayer | TextLayer>(selectedRecordKey);
-  const hasUIRecord = uiRecord != null;
-
-  const isShapeLayer = ShapeLayer.isModel(uiRecord);
-
-  const canEdit = hasUIRecord && isShapeLayer;
-
-  const strokeColor = isShapeLayer ? uiRecord.stroke.color : '';
-  const strokePattern = isShapeLayer ? uiRecord.stroke.pattern : '';
-  const strokeWidth = isShapeLayer ? uiRecord.stroke.width : { top: 0, right: 0, bottom: 0, left: 0 };
-
-  const onStrokeColorChange = useEvent((event: React.ChangeEvent<HTMLInputElement>, color: string) => {
-    if (!canEdit) {
-      return;
-    }
-    uiController.set<ShapeLayer>(uiRecord.key, { stroke: { color } });
-  });
-
-  const onStrokePatternChange = useEvent((event: React.ChangeEvent<HTMLSelectElement>, selected: string | number | undefined) => {
-    if (!canEdit || selected == null) {
-      return;
-    }
-    const pattern = selected as StrokeStylePattern;
-    uiController.set<ShapeLayer>(uiRecord.key, { stroke: { pattern } });
-  });
-
-  const onStrokeTopWidthChange = useEvent((event: React.ChangeEvent<HTMLInputElement>, value: number | null) => {
-    if (!canEdit) {
-      return;
-    }
-    const top = value ?? 0;
-    uiController.set<ShapeLayer>(uiRecord.key, { stroke: { width: { top } } });
-  });
-
-  const onStrokeRightWidthChange = useEvent((event: React.ChangeEvent<HTMLInputElement>, value: number | null) => {
-    if (!canEdit) {
-      return;
-    }
-    const right = value ?? 0;
-    uiController.set<ShapeLayer>(uiRecord.key, { stroke: { width: { right } } });
-  });
-
-  const onStrokeBottomWidthChange = useEvent((event: React.ChangeEvent<HTMLInputElement>, value: number | null) => {
-    if (!canEdit) {
-      return;
-    }
-    const bottom = value ?? 0;
-    uiController.set<ShapeLayer>(uiRecord.key, { stroke: { width: { bottom } } });
-  });
-
-  const onStrokeLeftWidthChange = useEvent((event: React.ChangeEvent<HTMLInputElement>, value: number | null) => {
-    if (!canEdit) {
-      return;
-    }
-    const left = value ?? 0;
-    uiController.set<ShapeLayer>(uiRecord.key, { stroke: { width: { left } } });
-  });
+  const {
+    canEdit,
+    strokeColor,
+    strokePattern,
+    strokeWidth,
+    onStrokeColorChange,
+    onStrokePatternChange,
+    onStrokeTopWidthChange,
+    onStrokeRightWidthChange,
+    onStrokeBottomWidthChange,
+    onStrokeLeftWidthChange,
+  } = viewModel;
+  const { selected, className, ...restProps } = props;
 
   return (
-    <Panel.Group as="div" {...props} ref={ref} className={clsx(styles.root, props.className)} heading="Stroke">
+    <Panel.Group as="div" {...restProps} ref={ref} className={clsx(styles.root, className)} heading="Stroke">
       <LayerFieldset>
         <LayerField label="Color">
           <ColorPicker value={strokeColor} onChange={onStrokeColorChange} disabled={!canEdit} />
