+++ apps/web/components/LayerTypographySection/LayerTypographySection.tsx
@@ -1,7 +1,5 @@
 import ColorPicker from '@pigyuma/design-system/components/ColorPicker';
 import TextArea from '@pigyuma/design-system/components/TextArea';
-import { useEvent } from '@pigyuma/react-utils';
-import { Artboard, ShapeLayer, TextLayer, useUIController, useUIRecord } from '@pigyuma/ui-design-tool';
 import clsx from 'clsx';
 import React from 'react';
 import LayerField from '../LayerField';
@@ -9,35 +7,16 @@ import LayerFieldset from '../LayerFieldset';
 import Panel from '../Panel';
 import * as styles from './LayerTypographySection.css';
 import { LayerTypographySectionProps, LayerTypographySectionRef } from './types';
+import useLayerTypographySection from './useLayerTypographySection';
 
 const LayerTypographySection = React.forwardRef<LayerTypographySectionRef, LayerTypographySectionProps>((props, ref) => {
-  const { selected: selectedRecordKey, className, ...restProps } = props;
+  const viewModel = useLayerTypographySection(props, ref);
+  if (viewModel == null) {
+    return null;
+  }
 
-  const uiController = useUIController();
-
-  const uiRecord = useUIRecord<Artboard | ShapeLayer | TextLayer>(selectedRecordKey);
-  const hasUIRecord = uiRecord != null;
-
-  const isTextLayer = TextLayer.isModel(uiRecord);
-
-  const canEdit = hasUIRecord && isTextLayer;
-
-  const textColor = isTextLayer ? uiRecord.textColor.color : '';
-  const content = isTextLayer ? uiRecord.content : '';
-
-  const onTextColorChange = useEvent((event: React.ChangeEvent<HTMLInputElement>, color: string) => {
-    if (!canEdit) {
-      return;
-    }
-    uiController.set<TextLayer>(uiRecord.key, { textColor: { color } });
-  });
-
-  const onContentChange = useEvent((event: React.ChangeEvent<HTMLTextAreaElement>, content: string) => {
-    if (!canEdit) {
-      return;
-    }
-    uiController.set<TextLayer>(uiRecord.key, { content });
-  });
+  const { canEdit, textColor, content, onTextColorChange, onContentChange } = viewModel;
+  const { selected, className, ...restProps } = props;
 
   return (
     <Panel.Group as="div" {...restProps} ref={ref} className={clsx(styles.root, className)} heading="Typography">
